修改一些配置

+++ SDWebImage4.X源码阅读201704/SDWebImage4.X源码阅读201704/MainViewController.m
@@ -30,7 +30,7 @@
 - (IBAction)getImageType:(id)sender {
     NSData *imageData = [NSData dataWithContentsOfFile:[[NSBundle mainBundle] pathForResource:@"rock.gif" ofType:nil]];
     SDImageFormat formate = [NSData sd_imageFormatForImageData:imageData];
-    NSString *message = [NSString stringWithFormat:@"%d",formate];
+    NSString *message = [NSString stringWithFormat:@"%ld",(long)formate];
     showMessage(message,self);
 }
